[MS] Refactoring of parsec module and added env checks
