+++ client/src/parsec/claim_device.ts
@@ -23,6 +23,7 @@ import {
 } from '@/parsec/types';
 
 import { DEFAULT_HANDLE, MOCK_WAITING_TIME, getClientConfig, wait } from '@/parsec/internals';
+import { needsMocks } from '@/parsec/environment';
 
 export class DeviceClaim {
   handle: Handle | null;
@@ -44,10 +45,10 @@ export class DeviceClaim {
   }
 
   async abort(): Promise<void> {
-    if (this.canceller !== null && window.isDesktop()) {
+    if (this.canceller !== null && !needsMocks()) {
       libparsec.cancel(this.canceller);
     }
-    if (this.handle !== null && window.isDesktop()) {
+    if (this.handle !== null && !needsMocks()) {
       libparsec.claimerGreeterAbortOperation(this.handle);
     }
     this.canceller = null;
@@ -80,7 +81,7 @@ export class DeviceClaim {
 
     this._assertState(true, true);
 
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       const clientConfig = getClientConfig();
       const result = await libparsec.claimerRetrieveInfo(clientConfig, eventCallback, invitationLink);
       if (result.ok) {
@@ -96,25 +97,23 @@ export class DeviceClaim {
         // cspell:disable-next-line
         label: 'Gale Dekarios',
       };
-      return new Promise<Result<UserOrDeviceClaimInitialInfoDevice, ClaimerRetrieveInfoError>>((resolve, _reject) => {
-        resolve({ok: true, value: {
-          tag: 'Device',
-          handle: DEFAULT_HANDLE,
-          greeterUserId: '1234',
-          greeterHumanHandle: {
-            email: 'gale@waterdeep.faerun',
-            // cspell:disable-next-line
-            label: 'Gale Dekarios',
-          },
-        }});
-      });
+      return {ok: true, value: {
+        tag: 'Device',
+        handle: DEFAULT_HANDLE,
+        greeterUserId: '1234',
+        greeterHumanHandle: {
+          email: 'gale@waterdeep.faerun',
+          // cspell:disable-next-line
+          label: 'Gale Dekarios',
+        },
+      }};
     }
   }
 
   async initialWaitHost():
   Promise<Result<DeviceClaimInProgress1Info, ClaimInProgressError>> {
     this._assertState(true, false);
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       this.canceller = await libparsec.newCanceller();
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       const result = await libparsec.claimerDeviceInitialDoWaitPeer(this.canceller, this.handle!);
@@ -128,20 +127,18 @@ export class DeviceClaim {
     } else {
       this.SASCodeChoices = ['5MNO', '6PQR', '7STU', '8VWX'];
       this.correctSASCode = '7STU';
-      return new Promise<Result<DeviceClaimInProgress1Info, ClaimInProgressError>>((resolve, _reject) => {
-        resolve({ok: true, value: {
-          handle: DEFAULT_HANDLE,
-          greeterSas: this.correctSASCode,
-          greeterSasChoices: this.SASCodeChoices,
-        }});
-      });
+      return {ok: true, value: {
+        handle: DEFAULT_HANDLE,
+        greeterSas: this.correctSASCode,
+        greeterSasChoices: this.SASCodeChoices,
+      }};
     }
   }
 
   async signifyTrust():
   Promise<Result<DeviceClaimInProgress2Info, ClaimInProgressError>> {
     this._assertState(true, false);
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       this.canceller = await libparsec.newCanceller();
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       const result = await libparsec.claimerDeviceInProgress1DoSignifyTrust(this.canceller, this.handle!);
@@ -153,19 +150,17 @@ export class DeviceClaim {
       return result;
     } else {
       this.guestSASCode = '1337';
-      return new Promise<Result<DeviceClaimInProgress2Info, ClaimInProgressError>>((resolve, _reject) => {
-        resolve({ok: true, value: {
-          handle: DEFAULT_HANDLE,
-          claimerSas: '1337',
-        }});
-      });
+      return {ok: true, value: {
+        handle: DEFAULT_HANDLE,
+        claimerSas: '1337',
+      }};
     }
   }
 
   async waitHostTrust():
   Promise<Result<DeviceClaimInProgress3Info, ClaimInProgressError>> {
     this._assertState(true, false);
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       this.canceller = await libparsec.newCanceller();
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       const result = await libparsec.claimerDeviceInProgress2DoWaitPeerTrust(this.canceller, this.handle!);
@@ -176,18 +171,16 @@ export class DeviceClaim {
       return result;
     } else {
       await wait(MOCK_WAITING_TIME);
-      return new Promise<Result<DeviceClaimInProgress3Info, ClaimInProgressError>>((resolve, _reject) => {
-        resolve({ok: true, value: {
-          handle: DEFAULT_HANDLE,
-        }});
-      });
+      return {ok: true, value: {
+        handle: DEFAULT_HANDLE,
+      }};
     }
   }
 
   async doClaim(deviceLabel: DeviceLabel):
   Promise<Result<DeviceClaimFinalizeInfo, ClaimInProgressError>> {
     this._assertState(true, false);
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       this.canceller = await libparsec.newCanceller();
       const result = await libparsec.claimerDeviceInProgress3DoClaim(
         // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
@@ -200,18 +193,16 @@ export class DeviceClaim {
       return result;
     } else {
       await wait(MOCK_WAITING_TIME);
-      return new Promise<Result<DeviceClaimFinalizeInfo, ClaimInProgressError>>((resolve, _reject) => {
-        resolve({ok: true, value: {
-          handle: DEFAULT_HANDLE,
-        }});
-      });
+      return {ok: true, value: {
+        handle: DEFAULT_HANDLE,
+      }};
     }
   }
 
   async finalize(password: string):
   Promise<Result<AvailableDevice, ClaimInProgressError>> {
     this._assertState(true, false);
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       const result = await libparsec.claimerDeviceFinalizeSaveLocalDevice(this.handle!, {tag: 'Password', password: password});
       if (result.ok) {
@@ -221,21 +212,19 @@ export class DeviceClaim {
       return result;
     } else {
       this.handle = null;
-      return new Promise<Result<AvailableDevice, ClaimInProgressError>>((resolve, _reject) => {
-        this.device = {
-          keyFilePath: '/path',
-          organizationId: 'MyOrg',
-          deviceId: 'deviceid',
-          humanHandle: {
-            label: 'A',
-            email: 'a@b.c',
-          },
-          deviceLabel: 'a@b',
-          slug: 'slug',
-          ty: DeviceFileType.Password,
-        };
-        resolve({ok: true, value: this.device});
-      });
+      this.device = {
+        keyFilePath: '/path',
+        organizationId: 'MyOrg',
+        deviceId: 'deviceid',
+        humanHandle: {
+          label: 'A',
+          email: 'a@b.c',
+        },
+        deviceLabel: 'a@b',
+        slug: 'slug',
+        ty: DeviceFileType.Password,
+      };
+      return {ok: true, value: this.device};
     }
   }
 }
