+++ client/src/parsec/claim_user.ts
@@ -22,6 +22,7 @@ import {
 } from '@/parsec/types';
 
 import { DEFAULT_HANDLE, MOCK_WAITING_TIME, getClientConfig, wait } from '@/parsec/internals';
+import { needsMocks } from '@/parsec/environment';
 
 export class UserClaim {
   handle: Handle | null;
@@ -43,10 +44,10 @@ export class UserClaim {
   }
 
   async abort(): Promise<void> {
-    if (this.canceller !== null && window.isDesktop()) {
+    if (this.canceller !== null && !needsMocks()) {
       libparsec.cancel(this.canceller);
     }
-    if (this.handle !== null && window.isDesktop()) {
+    if (this.handle !== null && !needsMocks()) {
       libparsec.claimerGreeterAbortOperation(this.handle);
     }
     this.canceller = null;
@@ -79,7 +80,7 @@ export class UserClaim {
 
     this._assertState(true, true);
 
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       const clientConfig = getClientConfig();
       const result = await libparsec.claimerRetrieveInfo(clientConfig, eventCallback, invitationLink);
       if (result.ok) {
@@ -95,26 +96,24 @@ export class UserClaim {
         // cspell:disable-next-line
         label: 'Gale Dekarios',
       };
-      return new Promise<Result<UserOrDeviceClaimInitialInfoUser, ClaimerRetrieveInfoError>>((resolve, _reject) => {
-        resolve({ok: true, value: {
-          tag: 'User',
-          handle: DEFAULT_HANDLE,
-          claimerEmail: 'shadowheart@swordcoast.faerun',
-          greeterUserId: '1234',
-          greeterHumanHandle: {
-            email: 'gale@waterdeep.faerun',
-            // cspell:disable-next-line
-            label: 'Gale Dekarios',
-          },
-        }});
-      });
+      return {ok: true, value: {
+        tag: 'User',
+        handle: DEFAULT_HANDLE,
+        claimerEmail: 'shadowheart@swordcoast.faerun',
+        greeterUserId: '1234',
+        greeterHumanHandle: {
+          email: 'gale@waterdeep.faerun',
+          // cspell:disable-next-line
+          label: 'Gale Dekarios',
+        },
+      }};
     }
   }
 
   async initialWaitHost():
   Promise<Result<UserClaimInProgress1Info, ClaimInProgressError>> {
     this._assertState(true, false);
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       this.canceller = await libparsec.newCanceller();
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       const result = await libparsec.claimerUserInitialDoWaitPeer(this.canceller, this.handle!);
@@ -128,20 +127,18 @@ export class UserClaim {
     } else {
       this.SASCodeChoices = ['1ABC', '2DEF', '3GHI', '4JKL'];
       this.correctSASCode = '2DEF';
-      return new Promise<Result<UserClaimInProgress1Info, ClaimInProgressError>>((resolve, _reject) => {
-        resolve({ok: true, value: {
-          handle: DEFAULT_HANDLE,
-          greeterSas: '2DEF',
-          greeterSasChoices: ['1ABC', '2DEF', '3GHI', '4JKL'],
-        }});
-      });
+      return {ok: true, value: {
+        handle: DEFAULT_HANDLE,
+        greeterSas: '2DEF',
+        greeterSasChoices: ['1ABC', '2DEF', '3GHI', '4JKL'],
+      }};
     }
   }
 
   async signifyTrust():
   Promise<Result<UserClaimInProgress2Info, ClaimInProgressError>> {
     this._assertState(true, false);
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       this.canceller = await libparsec.newCanceller();
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       const result = await libparsec.claimerUserInProgress1DoSignifyTrust(this.canceller, this.handle!);
@@ -153,19 +150,17 @@ export class UserClaim {
       return result;
     } else {
       this.guestSASCode = '1337';
-      return new Promise<Result<UserClaimInProgress2Info, ClaimInProgressError>>((resolve, _reject) => {
-        resolve({ok: true, value: {
-          handle: DEFAULT_HANDLE,
-          claimerSas: '1337',
-        }});
-      });
+      return {ok: true, value: {
+        handle: DEFAULT_HANDLE,
+        claimerSas: '1337',
+      }};
     }
   }
 
   async waitHostTrust():
   Promise<Result<UserClaimInProgress3Info, ClaimInProgressError>> {
     this._assertState(true, false);
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       this.canceller = await libparsec.newCanceller();
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       const result = await libparsec.claimerUserInProgress2DoWaitPeerTrust(this.canceller, this.handle!);
@@ -176,18 +171,16 @@ export class UserClaim {
       return result;
     } else {
       await wait(MOCK_WAITING_TIME);
-      return new Promise<Result<UserClaimInProgress3Info, ClaimInProgressError>>((resolve, _reject) => {
-        resolve({ok: true, value: {
-          handle: DEFAULT_HANDLE,
-        }});
-      });
+      return {ok: true, value: {
+        handle: DEFAULT_HANDLE,
+      }};
     }
   }
 
   async doClaim(deviceLabel: string, userName: string, email: string):
   Promise<Result<UserClaimFinalizeInfo, ClaimInProgressError>> {
     this._assertState(true, false);
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       this.canceller = await libparsec.newCanceller();
       const result = await libparsec.claimerUserInProgress3DoClaim(
         // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
@@ -200,18 +193,16 @@ export class UserClaim {
       return result;
     } else {
       await wait(MOCK_WAITING_TIME);
-      return new Promise<Result<UserClaimFinalizeInfo, ClaimInProgressError>>((resolve, _reject) => {
-        resolve({ok: true, value: {
-          handle: DEFAULT_HANDLE,
-        }});
-      });
+      return {ok: true, value: {
+        handle: DEFAULT_HANDLE,
+      }};
     }
   }
 
   async finalize(password: string):
   Promise<Result<AvailableDevice, ClaimInProgressError>> {
     this._assertState(true, false);
-    if (window.isDesktop()) {
+    if (!needsMocks()) {
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       const result = await libparsec.claimerUserFinalizeSaveLocalDevice(this.handle!, {tag: 'Password', password: password});
       if (result.ok) {
@@ -221,21 +212,19 @@ export class UserClaim {
       return result;
     } else {
       this.handle = null;
-      return new Promise<Result<AvailableDevice, ClaimInProgressError>>((resolve, _reject) => {
-        this.device = {
-          keyFilePath: '/path',
-          organizationId: 'MyOrg',
-          deviceId: 'deviceid',
-          humanHandle: {
-            label: 'A',
-            email: 'a@b.c',
-          },
-          deviceLabel: 'a@b',
-          slug: 'slug',
-          ty: DeviceFileType.Password,
-        };
-        resolve({ok: true, value: this.device});
-      });
+      this.device = {
+        keyFilePath: '/path',
+        organizationId: 'MyOrg',
+        deviceId: 'deviceid',
+        humanHandle: {
+          label: 'A',
+          email: 'a@b.c',
+        },
+        deviceLabel: 'a@b',
+        slug: 'slug',
+        ty: DeviceFileType.Password,
+      };
+      return {ok: true, value: this.device};
     }
   }
 }
