+++ client/src/parsec/index.ts
@@ -1,8 +1,6 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
 
 export * from '@/parsec/types';
-export * from '@/parsec/functions';
-export * from '@/parsec/client_information';
 export * from '@/parsec/claim_user';
 export * from '@/parsec/claim_device';
 export * from '@/parsec/greet_user';
@@ -11,3 +9,7 @@ export * from '@/parsec/validation';
 export * from '@/parsec/workspace';
 export * from '@/parsec/file';
 export * from '@/parsec/environment';
+export * from '@/parsec/invitation';
+export * from '@/parsec/login';
+export * from '@/parsec/user';
+export * from '@/parsec/organization';
